Move GraphListSource into pipeline.util.
